its more and more cut
